added rv which transform func which uses rand into random variable

+++ OmegaCore/src/ids/vectorid.jl
@@ -1,6 +1,6 @@
 
 # const ID = Int
-const VectorID = Vector{Int}
+const VectorID = Vector{Int64}
 
 @inline base(::Type{Vector{T}}) where T = T[]
